requirements.txt
weasyprint
pybars3
pyyaml